requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.7
ollama==0.1.7 